backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
httpx>=0.27
pydantic>=2.8
mcp[cli]>=1.2
PyMuPDF>=1.24
python-docx>=1.1
rapidfuzz>=3.9
